chore: rename hook name
