+++ apps/web/src/pages/public-workspace/[workspaceId]/index.tsx
@@ -1,6 +1,6 @@
 import { PageList } from '@/components/page-list';
 import { WorkspaceUnitAvatar } from '@/components/workspace-avatar';
-import { useLoadPublicWorkspace } from '@/hooks/use-public-workspace';
+import { useLoadPublicWorkspace } from '@/hooks/use-load-public-workspace';
 import { PageMeta } from '@/providers/app-state-provider';
 import { useModal } from '@/store/globalModal';
 import { Breadcrumbs } from '@affine/component';
